Handle component paths > 63 characters
Component paths longer than 63 characters can't be used as field names in `Property`s.

To handle this, we introduce a hash value for the type_paths over 63 chars.
This hash value is *only* used as the key of the ComponentContainer.
This means we limit the scope of impact to accessing the root component "variant".
If we can, in the future, make this a getter on the ComponentContainer class, it would help clean up the code.
For now we just have to remember to use the "potentially hashed type_path" as the key when doing ComponentContainer access

This is a backwards-compatible change as previously these type_paths were not constructible, so no user will have data stored like this.

This change specifically enables

* `bevy_core_pipeline::contrast_adaptive_sharpening::ContrastAdaptiveSharpening`
* multiple Avian types, like `avian3d::dynamics::rigid_body::mass_properties::components::Mass`

Unfortunately, this means words like

`SKEIN_E96B456420F7FAB0B40CCCBD30A331F3` shows in the UI because `avian3d::dynamics::rigid_body::mass_properties::components::Mass` is a tuple type. We can improve that in the future by using "shortName"s maybe for display purposes. tbd